+++ src/bump_vec.rs
@@ -6,7 +6,7 @@ use crate::{
     collection_method_allocator_stats,
     destructure::destructure,
     error_behavior_generic_methods_allocation_failure, min_non_zero_cap,
-    owned_slice::{self, OwnedSlice},
+    owned_slice::{self, IntoOwnedSlice, OwnedSlice},
     polyfill::{nonnull, pointer, slice},
     raw_bump_box::RawBumpBox,
     raw_fixed_bump_vec::RawFixedBumpVec,
@@ -1377,9 +1377,9 @@ impl<T, A: BumpAllocator> BumpVec<T, A> {
         /// ```
         for fn try_append
         #[inline]
-        use fn generic_append(&mut self, other: impl OwnedSlice<Item = T>) {
+        use fn generic_append(&mut self, other: impl IntoOwnedSlice<Item = T>) {
             unsafe {
-                let mut owned_slice = other;
+                let mut owned_slice = other.into_owned_slice();
 
                 let slice = owned_slice.owned_slice_ptr();
                 self.generic_reserve(slice.len())?;