+++ src/fixed_bump_vec.rs
@@ -1,6 +1,6 @@
 use crate::{
-    error_behavior_generic_methods_allocation_failure, error_behavior_generic_methods_if, owned_slice,
-    owned_slice::OwnedSlice,
+    error_behavior_generic_methods_allocation_failure, error_behavior_generic_methods_if,
+    owned_slice::{self, IntoOwnedSlice, OwnedSlice},
     polyfill::{self, nonnull, pointer, slice},
     set_len_on_drop_by_ptr::SetLenOnDropByPtr,
     BumpAllocatorScope, BumpBox, BumpVec, ErrorBehavior, NoDrop, SizedTypeProperties,
@@ -1009,9 +1009,9 @@ impl<'a, T> FixedBumpVec<'a, T> {
         /// ```
         for fn try_append
         #[inline]
-        use fn generic_append(&mut self, other: impl OwnedSlice<Item = T>) {
+        use fn generic_append(&mut self, other: impl IntoOwnedSlice<Item = T>) {
             unsafe {
-                let mut owned_slice = other;
+                let mut owned_slice = other.into_owned_slice();
 
                 let slice = owned_slice.owned_slice_ptr();
                 self.generic_reserve(slice.len())?;